admin_api/requirements.txt
amqp==5.1.1
asgiref==3.5.0
billiard==4.1.0
celery==5.3.1
certifi==2021.10.8
charset-normalizer==2.0.11
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
Django==3.2.12
django-stubs==4.2.3
django-stubs-ext==4.2.2
djangorestframework==3.13.1
idna==3.3
kombu==5.3.1
mypy==1.4.1
mypy-extensions==1.0.0
prompt-toolkit==3.0.39
psycopg2==2.9.3
python-dateutil==2.8.2
pytz==2021.3
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
types-pytz==2023.3.0.0
types-PyYAML==6.0.12.11
typing_extensions==4.7.1
tzdata==2023.3
urllib3==1.26.8
vine==5.0.0
wcwidth==0.2.6
